+++ src/main/java/plugily/projects/buildbattle/commands/TabCompletion.java
@@ -18,6 +18,8 @@
 
 package plugily.projects.buildbattle.commands;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -25,8 +27,10 @@ import java.util.stream.Collectors;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.TabCompleter;
-import org.bukkit.entity.Player;
+import org.bukkit.configuration.file.FileConfiguration;
+import org.bukkit.util.StringUtil;
 
+import pl.plajerlair.commonsbox.minecraft.configuration.ConfigUtils;
 import plugily.projects.buildbattle.arena.ArenaRegistry;
 import plugily.projects.buildbattle.arena.impl.BaseArena;
 import plugily.projects.buildbattle.commands.arguments.ArgumentsRegistry;
@@ -47,20 +51,61 @@ public class TabCompletion implements TabCompleter {
 
   @Override
   public List<String> onTabComplete(CommandSender sender, Command cmd, String label, String[] args) {
-    if (!(sender instanceof Player)) {
-      return Collections.emptyList();
-    }
-    if (cmd.getName().equalsIgnoreCase("buildbattleadmin") && args.length == 1) {
-      return registry.getMappedArguments().get(cmd.getName().toLowerCase()).stream().map(CommandArgument::getArgumentName).collect(Collectors.toList());
-    }
-    if (cmd.getName().equalsIgnoreCase("buildbattle")) {
-      if (args.length == 2 && args[0].equalsIgnoreCase("join")) {
-        return ArenaRegistry.getArenas().stream().map(BaseArena::getID).collect(Collectors.toList());
-      }
+    List<String> completionList = new ArrayList<>(), cmds = new ArrayList<>();
+    String partOfCommand = null;
+
+    if (cmd.getName().equalsIgnoreCase("buildbattleadmin")) {
       if (args.length == 1) {
-        return registry.getMappedArguments().get(cmd.getName().toLowerCase()).stream().map(CommandArgument::getArgumentName).collect(Collectors.toList());
+        cmds.addAll(registry.getMappedArguments().get(cmd.getName().toLowerCase()).stream().map(CommandArgument::getArgumentName)
+            .collect(Collectors.toList()));
+        partOfCommand = args[0];
+      } else if (args.length == 2) {
+        if (args[0].equalsIgnoreCase("delete") || args[0].equalsIgnoreCase("removeplot")
+          || args[0].equalsIgnoreCase("addnpc") || args[0].equalsIgnoreCase("addplot")) {
+            cmds.addAll(ArenaRegistry.getArenas().stream().map(BaseArena::getID).collect(Collectors.toList()));
+        } else if (args[0].equalsIgnoreCase("settheme")) {
+          for (List<String> l : registry.getPlugin().getConfigPreferences().getGameThemes().values()) {
+            cmds.addAll(l);
+          }
+        } else if (args[0].equalsIgnoreCase("votes")) {
+          Arrays.asList("add", "set").forEach(cmds::add);
+        }
+
+        partOfCommand = args[1];
+      } else if (args.length == 3 && args[0].equalsIgnoreCase("removeplot")) {
+        FileConfiguration config = ConfigUtils.getConfig(registry.getPlugin(), "arenas");
+        String path = "instances." + ArenaRegistry.getArena(args[1]).getID() + ".plots";
+        cmds.addAll(config.isConfigurationSection(path) ? config.getConfigurationSection(path)
+            .getKeys(false).stream().collect(Collectors.toList()) : Collections.emptyList());
+        partOfCommand = args[2];
       }
     }
-    return Collections.emptyList();
+
+    if (cmd.getName().equalsIgnoreCase("buildbattle")) {
+      if (args.length >= 2 && (args[0].equalsIgnoreCase("join") || args[0].equalsIgnoreCase("randomjoin"))) {
+        if (args.length == 3 && args[0].equalsIgnoreCase("randomjoin")) {
+          Arrays.asList("solo", "team", "gtb", "guess_the_build").forEach(cmds::add);
+          partOfCommand = args[2];
+        } else {
+          cmds.addAll(ArenaRegistry.getArenas().stream().map(BaseArena::getID).collect(Collectors.toList()));
+          partOfCommand = args[1];
+        }
+      } else if (args.length == 1) {
+        cmds.addAll(registry.getMappedArguments().get(cmd.getName().toLowerCase()).stream().map(CommandArgument::getArgumentName)
+            .collect(Collectors.toList()));
+        partOfCommand = args[0];
+      } else if (args.length == 2 && args[0].equalsIgnoreCase("votes")) {
+        Arrays.asList("add", "set").forEach(cmds::add);
+      }
+    }
+
+    // Completes the player names
+    if (cmds.isEmpty() || partOfCommand == null) {
+      return null;
+    }
+
+    StringUtil.copyPartialMatches(partOfCommand, cmds, completionList);
+    Collections.sort(completionList);
+    return completionList;
   }
 }