+++ CHANGELOG.md
@@ -11,6 +11,7 @@ That's all, matcher will stop when detects next line started with `###` match
 
 ### 4.4.1-b5 Beta
 * Fix NoSuchMethodException for player chunk map
+* Improved tab complete handling
 
 ### 4.4.1-b4 Beta
 * Fix for NPE when teleporting players to plots